pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "recsys_tfb"
version = "0.1.0"
description = "Product recommendation ranking model for a commercial bank"
requires-python = ">=3.10"
dependencies = [
    "pyspark>=3.3.2",
    "pyyaml>=6.0",
    "pandas>=1.5.3",
    "numpy>=1.25.0",
    "pyarrow>=14.0.1",
    "lightgbm>=4.6.0",
    "scikit-learn>=1.5.0",
    "mlflow>=3.1.0",
    "optuna>=4.5.0",
    "shap>=0.42.1",
]

[project.optional-dependencies]
dev = [
    "pytest==7.3.1",
    "pytest-cov>=4.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
